+++ src/index.ts
@@ -22,6 +22,7 @@ export { default as GanttSlack } from './components/GanttSlack.vue'
 export { default as GanttDeadlines } from './components/GanttDeadlines.vue'
 export { default as GanttBaselines } from './components/GanttBaselines.vue'
 export { default as GanttPeriods } from './components/GanttPeriods.vue'
+export { default as GanttMarkers } from './components/GanttMarkers.vue'
 export { default as GanttToday } from './components/GanttToday.vue'
 export { default as GanttZoom } from './components/GanttZoom.vue'
 
@@ -125,12 +126,14 @@ export type {
   GanttItemType,
   GanttLinkDraft,
   GanttLinkMode,
+  GanttMarker,
   GanttMoveEvent,
   GanttOverlapMode,
   GanttPeriod,
   GanttProgressEvent,
   GanttRangeChangeEvent,
   GanttResizeEvent,
+  GanttResource,
   GanttRowEditEvent,
   GanttTaskEditEvent,
   GanttRootProps,
@@ -148,8 +151,10 @@ export type {
   NonWorkingBand,
   NonWorkingCalendar,
   ResolvedGroup,
+  ResolvedMarker,
   ResolvedNonWorkingBand,
   ResolvedPeriod,
+  ResolvedResource,
   ResolvedRow,
   ResolvedSegment,
   ResolvedTask,